Update MotifSearch.java
In this class project I used Java Language. The program searches all the possible combinations of four letters A, C, G, and T. In which there are 256 different combinations. It takes three different inputs of length ten and puts them into seven possible substrings and attempts to match them with all 256 combinations. It takes the three inputs and matches the substrings from all three and then displays the output.